# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Mastori Api specification",
        "description": "",
        "contact": {
            "email": "giannaropoulou.foteini@gmail.com;tania.pets@gmail.com"
        },
        "version": "1.0.0"
    },
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/mastoria": {
            "get": {
                "tags": [
                    "mastoria"
                ],
                "description": "Returns all the mastoria",
                "operationId": "getMastoria",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Access token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All mastoria response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/mastori"
                            }
                        }
                    },
                    "400": {
                        "description": "token_not_provided",
                        "schema": {
                            "$ref": "#/definitions/errorModel"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "mastoria"
                ],
                "description": "Adds a new mastori in database",
                "operationId": "addMastori",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mastoriPost"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Returns created Mastori object",
                        "schema": {
                            "$ref": "#/definitions/mastori"
                        }
                    },
                    "400": {
                        "description": "validation errors",
                        "schema": {
                            "$ref": "#/definitions/validationsErrorsModel"
                        }
                    }
                }
            }
        },
        "/mastoria/{id}": {
            "get": {
                "tags": [
                    "mastoria"
                ],
                "description": "Fetches a mastori",
                "operationId": "getmastori",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of mastori to get",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Access token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mastori response",
                        "schema": {
                            "$ref": "#/definitions/mastori"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/errorModel"
                        }
                    },
                    "400": {
                        "description": "token_not_provided/token_invalid",
                        "schema": {
                            "$ref": "#/definitions/errorModel"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "mastoria"
                ],
                "description": "Edits a mastori",
                "operationId": "editmastori",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of mastori to update",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mastori"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mastori response",
                        "schema": {
                            "$ref": "#/definitions/mastori"
                        }
                    },
                    "401": {
                        "description": "token_not_provided/token_invalid",
                        "schema": {
                            "$ref": "#/definitions/errorModel"
                        }
                    },
                    "400": {
                        "description": "validation errors",
                        "schema": {
                            "$ref": "#/definitions/validationsErrorsModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "address": {
            "required": [
                "lat",
                "lng",
                "address",
                "city",
                "country"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "readOnly": true
                },
                "lat": {
                    "type": "number",
                    "format": "double"
                },
                "lng": {
                    "type": "number",
                    "format": "double"
                },
                "address": {
                    "type": "string"
                },
                "friendly_name": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                }
            }
        },
...
        "errorModel": {
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "type": "string",
                    "readOnly": true
                }
            }
        },
        "fieldValidationDescription": {
            "properties": {
                "error_description": {
                    "type": "string"
                }
            }
        },
        "validationerror": {
            "properties": {
                "fieldName": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fieldValidationDescription"
                    }
                }
            }
        },
        "validationsErrorsModel": {
            "required": [
                "errors"
            ],
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/validationerror"
                    }
                }
            }
        },
        "mastori": {
            "required": [
                "username",
                "email",
                "password",
                "last_name",
                "first_name",
                "phone",
                "pricelist"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "readOnly": true
                },
                "username": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "format": "password"
                },
                "last_name": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "paratsoukli": {
                    "type": "string"
                },
                "facebook_id": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "pricelist": {
                    "type": "string"
                },
                "avg_response_time": {
                    "type": "integer"
                },
                "avg_rating": {
                    "type": "number",
                    "format": "float"
                },
                "active": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                },
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/address"
                    }
                },
                "professions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/profession"
                    }
                }
            },
            "type": "object"
        },
        "mastoriPost": {
            "properties": {
                "password_repeat": {
                    "type": "string"
                }
            },
            "allOf": [
                {
                    "$ref": "#definitions/mastori"
                }
            ]
        },
...
        "profession": {
            "required": [
                "tag",
                "title"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "readOnly": true
                },
                "tag": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                }
            }
        },
...
    }
}